fix: 검증 로직 이전
